Return page size as unsigned long
Currently page_size() returns unsigned int value that is after "bitwise
not" is promoted to unsigned long value e.g. in uffd.c
handle_page_fault. Since the value is unsigned promotion is done with 0
MSB that results in lost of MSB pagefault address bits. So make
page_size to return  unsigned long to avoid such situation.

+++ include/common/arch/ppc64/asm/page.h
@@ -10,7 +10,7 @@
 extern unsigned __page_size;
 extern unsigned __page_shift;
 
-static inline unsigned page_size(void)
+static inline unsigned long page_size(void)
 {
 	if (!__page_size)
 		__page_size = sysconf(_SC_PAGESIZE);
@@ -37,7 +37,7 @@ static inline unsigned page_shift(void)
 
 #else /* CR_NOGLIBC */
 
-extern unsigned page_size(void);
+extern unsigned long page_size(void);
 #define PAGE_SIZE page_size()
 
 #endif /* CR_NOGLIBC */
